role fk in accident_statement

+++ app/api/routes/accident.py
@@ -24,7 +24,7 @@ from app.db.repositories.temporary_accident_driver_data import TemporaryReposito
 from app.api.dependencies.database import get_repository
 from app.api.dependencies.auth import get_current_active_user
 from fastapi.responses import FileResponse, StreamingResponse
-import io
+import io, os
 
 router = APIRouter()
 
@@ -51,20 +51,16 @@ async def get_accident_by_id(id: int,
     if current_user.is_superuser:
         return await accident_repo.get_accident_by_id(id = id)
     else:
+        print("ok")
         accident = await accident_repo.get_accident_from_user_with_statement_id(id=id, user_id = current_user.id)
         accident1 = await accident_repo.get_accident_by_temporary_driver_by_email(id=id, email = current_user.email)
         if not accident:
             if not accident1: 
                 raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="No accident found")
-            else:
-                # if current_user.email == accident.temporary_accident_drivers[0].driver_email:         
+            else:       
                 accident = accident1
                 return accident
-                # else:
-                #     raise HTTPException(status_code=HTTP_401_UNAUTHORIZED, detail="No access to this accident")
         else:
-            # if current_user.id != accident.accident_statement[0].user_id:
-            #     raise HTTPException(status_code=HTTP_401_UNAUTHORIZED, detail="No access to this accident")
             return accident
 
 @router.post("/temporary/{accident_id}/",
@@ -83,23 +79,27 @@ async def add_other_drivers_by_accident_id(accident_id : int,
 
     email_list = []
     accident_statements =  await accidentstmt_repo.get_all_accident_statements_for_accident_id(accident_id=accident_id)
-    print(accident_statements)
     if not accident_statements:
         raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="No accident found")
     else:
         for accident_statement in accident_statements:
             email_list.append(accident_statement.user.email)
-    print(email_list)
     if new_temporary.driver_email in email_list:
-         raise HTTPException(status_code=HTTP_400_BAD_REQUEST, detail="Driver is already added")
+        raise HTTPException(status_code=HTTP_400_BAD_REQUEST, detail="Driver is already added")
     stmt_list = []
     for i in range(len(accident.accident_statement)):
         stmt_list.append(accident.accident_statement[i].id)
     accident_statement = await accidentstmt_repo.get_accident_statement_by_accident_id_user_id(accident_id = accident_id, user_id = current_user.id, populate = False)
+
     if not accident_statement:
-        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You have not made a statement")        
-    elif accident_statement.id != min(stmt_list):
-        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You can not remove drivers")
+        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You have not made a statement")       
+
+    if accident_statement.id != min(stmt_list):
+        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You can not add drivers")
+
+    if accident_statement.done:
+        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You have declares this accident statement as completed")  
+
     temporary_driver_data = await temporary_repo.add_driver_to_accident(new_temporary= new_temporary, accident_id = accident.id)
     return temporary_driver_data
 
@@ -109,16 +109,18 @@ async def get_accidents_by_id(
     current_user: UserPublic = Depends(get_current_active_user),
     accident_repo: AccidentRepository = Depends(get_repository(AccidentRepository)),
     ) -> List[AccidentPublic]:
-    if current_user.is_superuser:
+    if current_user.is_master:
         return await accident_repo.get_all_accidents()
+    elif current_user.is_superuser:
+        return await accident_repo.get_all_accidents_by_insurance_company(insurance_company_email=current_user.email)
     else:
-        accidents = await accident_repo.get_accidents_by_user_id(user_id = current_user.id, email = current_user.email)
+        accidents = await accident_repo.get_accidents_by_user_id(user_id = current_user.id, email=current_user.email)
         if not accidents:
             raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="No accident found")
         return accidents
 
 
-@router.post("/accident_stmt/{accident_id}")
+@router.post("/accident_stmt/{accident_id}",name="accident:add-accident-stmt")
 async def create_new_accident_statement(
     accident_id: int,
     current_user: UserPublic = Depends(get_current_active_user),
@@ -129,7 +131,6 @@ async def create_new_accident_statement(
     accident_stmt_repo: AccidentStatementRepository = Depends(get_repository(AccidentStatementRepository)),
     ) -> Accident_statement_Public:
     accident = await accident_repo.get_accident_by_id(id = accident_id)
-
     accident_permission = await temporary_repo.get_temporary_driver_data_by_driver_email(accident_id=accident_id, email=current_user.email)    
     print(accident_permission)
     vehicle = await vehicles_repo.get_vehicle_by_user_id_digit(sign = accident_permission['vehicle_sign'], user_id = current_user.id)
@@ -139,7 +140,9 @@ async def create_new_accident_statement(
             )
 
     new_accident_stmt.vehicle_id = vehicle.id 
+    new_accident_stmt.role_id= vehicle.roles.id
     new_accident_stmt.insurance_id = vehicle.insurance.id
+
     expiredate = datetime.datetime(
         year=vehicle.insurance.expire_date.year, 
         month=vehicle.insurance.expire_date.month,
@@ -166,7 +169,7 @@ async def create_new_accident_statement(
      
     return accident_stmt
 
-@router.put("/accident_stmt/{accident_id}")
+@router.put("/accident_stmt/{accident_id}",name="accident:edit-accident-stmt")
 async def update_accident_statement(
     accident_id: int,
     current_user: UserPublic = Depends(get_current_active_user),
@@ -179,10 +182,13 @@ async def update_accident_statement(
     if not accident_stmt:
         raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="Please can not update this stmt")
 
+    if accident_stmt.done:
+        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You have declared this accident statement as completed")  
+
     updated_stmt = await accident_stmt_repo.update_accident_statement(accident_id= accident_id, user_id = current_user.id, accident_statement_update = accident_statement_update)
     return updated_stmt
 
-@router.put("/accident_stmt_detection/{accident_id}")
+@router.put("/accident_stmt_detection/{accident_id}",name="accident:add-damaged-part-detected")
 async def update_accident_statement_detection(
     accident_id: int,
     current_user: UserPublic = Depends(get_current_active_user),
@@ -193,7 +199,10 @@ async def update_accident_statement_detection(
     accident_stmt = await accident_stmt_repo.get_accident_statement_by_accident_id_user_id(accident_id= accident_id, user_id = current_user.id)
 
     if not accident_stmt:
-        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="Please can not update this stmt")
+        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You can not update this stmt")
+
+    if accident_stmt.done:
+        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You have declared this accident statement as completed")  
 
     updated_stmt = await accident_stmt_repo.update_accident_statement_detection(accident_id= accident_id, user_id = current_user.id, accident_statement_update = accident_statement_update)
     return updated_stmt
@@ -211,6 +220,10 @@ async def create_new_accident_sketch(
     accident_stmt = await accident_stmt_repo.get_accident_statement_by_accident_id_user_id(accident_id= accident_id, user_id = current_user.id)
     if not accident_stmt:
         raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You can not update this sketch")
+
+    if accident_stmt.done:
+        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You have declared this accident statement as completed")  
+
     new_accident_sketch.statement_id=accident_stmt.id
     created_sketch = await sketch_repo.create_new_accident_sketch(new_accident_sketch=new_accident_sketch, statement_id= accident_stmt.id)
     return created_sketch
@@ -233,46 +246,65 @@ async def remove_accident_driver(
         stmt_list.append(accident.accident_statement[i].id)
     accident_statement = await accidentstmt_repo.get_accident_statement_by_accident_id_user_id(accident_id = accident.id, user_id = current_user.id, populate = False)
     if not accident_statement:
-        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You have not made a statement")        
-    elif accident_statement.id != min(stmt_list):
+        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You have not made a statement")
+
+    if accident_statement.done:
+        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You have declared this accident statement as completed")
+        
+    if accident_statement.id != min(stmt_list):
         raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You can not remove drivers")
     temporary_driver_data = await temporary_repo.delete_temporary_by_id(id = id)
     accident_upd = await accident_repo.get_accident_from_user_with_statement_id(id = accident.id, user_id = current_user.id, populate = True)
     return accident_upd
 
-@router.post("/newImage")
+@router.post("/newImage", name="accident:add-accident-image")
 async def create_new_accident_image(
  accident_id:  int = Form(...),
  image: UploadFile = File(...),
  current_user: UserPublic = Depends(get_current_active_user),
  accident_image_repo: AccidentImageRepository = Depends(get_repository(AccidentImageRepository)),
  accident_stmt_repo: AccidentStatementRepository = Depends(get_repository(AccidentStatementRepository)),
-    ) -> Accident_Image_Public:
+    ):
 
     accident_stmt = await accident_stmt_repo.get_accident_statement_by_accident_id_user_id(accident_id= accident_id, user_id = current_user.id)
     if not accident_stmt:
-        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You can not add image")
+        raise HTTPException(status_code=HTTP_400_BAD_REQUEST, detail="You can not add image")
+    print("OK")
+    if accident_stmt.done:
+        raise HTTPException(status_code=HTTP_400_BAD_REQUEST, detail="You have declared this accident statement as completed")  
+
     contentsImage = await image.read()
     new_accident_image = Accident_Image_Create(statement_id=accident_stmt.id, image = contentsImage)
     accident_image = await accident_image_repo.add_new_accident_image(new_accident_image = new_accident_image)
     return accident_image
 
-@router.get("/image/{accident_id}")
+@router.get("/image/{id}", name="accident:get-accident-image")
 async def get_image_from_accident_stmt(
-    accident_id:int,
+    id:int,
     current_user: UserPublic = Depends(get_current_active_user),
+    temporary_repo: TemporaryRepository = Depends(get_repository(TemporaryRepository)),
     accident_image_repo: AccidentImageRepository = Depends(get_repository(AccidentImageRepository)),
     accident_stmt_repo: AccidentStatementRepository = Depends(get_repository(AccidentStatementRepository)),
     ) -> bytes:
-    accident_stmt = await accident_stmt_repo.get_accident_statement_by_accident_id_user_id(accident_id= accident_id, user_id = current_user.id)
+    image_data = await accident_image_repo.get_statement(id = id)
+    print(image_data.statement_id)
+    if not image_data.statement_id:
+         raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You have not access to this accident")
+    accident_stmt = await accident_stmt_repo.get_accident_statement_by_id(id=image_data.statement_id)
+    print(accident_stmt)
     if not accident_stmt:
         raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You have not access to this accident")
-    image = await accident_image_repo.get_image(statement_id = accident_stmt.id)
-    # print(image.image)
-    # file_like = open(os.path.join(IMG_DIR, 'nikitas.jpg'), mode="rb")
+    temporary = await temporary_repo.get_all_temporary_driver_data_for_accident_id(accident_id=accident_stmt.accident_id)
+    email_list = []
+    for temporary_accident_driver in temporary:
+        email_list.append(temporary_accident_driver["driver_email"])
+    if current_user.is_master or current_user.is_superuser or current_user.id == accident_stmt.user_id or current_user.email in email_list:
+        image = await accident_image_repo.get_image(id = id)
+    else:
+        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You have not access to this accident")
     return StreamingResponse(io.BytesIO(image.image), media_type="image/jpeg")
 
-@router.get("/imageList/{accident_id}")
+@router.get("/imageList/{accident_id}", name="accident:get-accident-image-list")
 async def get_image_count_from_accident_stmt(
     accident_id:int,
     current_user: UserPublic = Depends(get_current_active_user),
@@ -284,6 +316,39 @@ async def get_image_count_from_accident_stmt(
         raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You have not access to this accident")
     ids = await accident_image_repo.get_image_count(statement_id = accident_stmt.id)
     return ids
+
+@router.put("/complete_statement/{accident_id}", response_model=Accident_statement_Public, name="accident_statement:complete-statement")
+async def complete_accident_statement(accident_id:int,
+    current_user: UserInDB = Depends(get_current_active_user),
+    accident_stmt_repo: AccidentStatementRepository = Depends(get_repository(AccidentStatementRepository)),
+    temporary_repo: TemporaryRepository = Depends(get_repository(TemporaryRepository)),
+    accident_image_repo: AccidentImageRepository = Depends(get_repository(AccidentImageRepository)),
+    sketch_repo: AccidentSketchRepository = Depends(get_repository(AccidentSketchRepository)),
+)-> Accident_statement_Public:
+    temporary = await temporary_repo.get_all_temporary_driver_data_for_accident_id(accident_id=accident_id)
+    if not temporary:
+        raise HTTPException(status_code=HTTP_400_BAD_REQUEST, detail="You should add another driver first")
+    accident_stmt = await accident_stmt_repo.get_accident_statement_by_accident_id_user_id(accident_id= accident_id, user_id = current_user.id)
+    if not accident_stmt:
+        raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="You have no access to this accident")
+    if accident_stmt.caused_by == 'add cause' or accident_stmt.comments == '' or accident_stmt.comments == 'add comment':
+        raise HTTPException(status_code=HTTP_400_BAD_REQUEST, detail="You have to complete the accident stmt")
+    image_count = await accident_image_repo.get_image_count(statement_id = accident_stmt.id)
+    email_list = []
+    for temporary_accident_driver in temporary:
+        email_list.append(temporary_accident_driver["driver_email"])
+    if image_count == 0 and current_user.email not in email_list:
+        raise HTTPException(status_code=HTTP_400_BAD_REQUEST, detail="You have to add at least one picture")
+
+    stmt_sketch = await sketch_repo.get_accident_sketch_by_statement_id(statement_id = accident_stmt.id)
+    if not stmt_sketch:
+        raise HTTPException(status_code=HTTP_400_BAD_REQUEST, detail="You have to draw the accident")
+    updated_stmt = await accident_stmt_repo.complete_accident_statement(accident_id= accident_id, user_id = current_user.id)
+    return updated_stmt
+    
+
+    
+
